Refactor OpalDefendantAccountHistoryService to use EnforcementHistorySourceService for fetching enforcements
